requirements.txt
streamlit
pandas
numpy
requests
python-dotenv
SQLAlchemy
plotly
psycopg2-binary

pydeck
